requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
drf-yasg==1.20.0
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1